refactor(nef): extract scanner error module
Move ScanError and its message helpers to scan/error.rs. Re-exported
from scan so the crate-public path crate::scan::ScanError is unchanged.
